Using documentById and switching over to regular restaurants still needs doing
